requirements.txt
aiogram>=3.6,<4
python-dotenv>=1.0
loguru>=0.7
fastapi>=0.111
uvicorn[standard]>=0.27
Jinja2>=3.1
